Criado 'createUser' na userApi
-- Criado um método 'createUser' na userApi que é responsável por criar o usuário. Além disso esse método automaticamente chema o create_accounts da currencyApi, não sendo necessário fazer isso automaticamente
-- Adaptado o register para usar o novo método da userApi
