+++ package.json
@@ -38,6 +38,7 @@
     "@types/express": "^4.17.2",
     "@types/mongoose": "^5.5.32",
     "@types/node": "^12.12.11",
+    "@types/randomstring": "^1.1.6",
     "@types/socket.io": "^2.1.4",
     "@types/socket.io-client": "^1.4.32",
     "@types/ws": "^6.0.3",
